# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAPhoto Backend
  description: |-
    Keep in touch with your friends by sharing photos of special moments, thanks to WASAPhoto! You can
    upload your photos directly from your PC, and they will be visible to everyone following you.
  contact:
    email: deodati.2025953@studenti.uniroma1.it
  license:
    name: GPL 3.0
    url: https://www.gnu.org/licenses/gpl-3.0.en.html
  version: 1.0.0
servers:
  - url: "http://localhost:3000"
...
paths:
  /session:
    post:
      tags:
        - Login
      summary: Logs in the user
      description: Attempts to login, if successul the user token is returned.
      operationId: doLogin
      requestBody:
        description: User details
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCredentials'
        required: true
      responses:
        '200':
          description: User log-in action successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessToken'
        '403':
          description: Access denied
  /users:
    get:
      tags:
        - Users
      summary: Search for users
      description: Search registered users on the website
      operationId: searchUsers
      parameters:
        - in: query
          name: q
          description: Your search query
          required: true
          schema:
            type: string
            minLength: 3
            maxLength: 255
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserList'
        '400':
          description: Empty query
    post:
      tags:
        - Login
      summary: Register to the website
      description: Registers a new user and issues an access token after successful registration
      operationId: register
      requestBody:
        description: Registration data
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegistrationRequest'
      responses:
        '200':
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessToken'
        '400':
          description: Weak password
        '409':
          description: Username already taken
...
  /users/me/followers:
    get:
      tags:
        - Followers
      summary: List followers
      description: See who is following you
      operationId: getFollowers
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ID'
        '400':
          description: Bad auth token
        '401':
          description: Unauthenticated
        '403':
          description: "Forbidden: cannot view somebody else's followers"
  /users/me/followers/{followerID}:
    delete:
      tags:
        - Followers
      summary: Remove follower
      description: Removes a follower from your followers' list
      operationId: removeFollower
      parameters:
        - in: path
          name: followerID
          description: Follower to remove
          required: true
          schema:
            $ref: '#/components/schemas/ID'
      responses:
        '204':
          description: Successful operation
        '400':
          description: Bad auth token or trying to remove yourself
        '401':
          description: Unauthenticated
        '403':
          description: Trying to edit somebody else's followers
        '404':
          description: User not found or not following
  /users/me/blocked:
    get:
      tags:
        - Followers
      summary: List blocked
      description: See who you've blocked
      operationId: getBlocked
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ID'
        '400':
          description: Bad auth token
        '401':
          description: Unauthenticated
        '403':
          description: "Forbidden: cannot view somebody else's blocks"
    post:
      tags:
        - Followers
      summary: Block user
      description: |-
        Blocks annoying user. On success returns the blocked user's ID.
        Also removes the user from your followers
      operationId: banUser
      requestBody:
        description: user ID to block
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ID'
      responses:
        '200':
          description: Successful operation
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ID'
        '400':
          description: Bad auth token, bad userID, already blocked or trying to block yourself
        '401':
          description: Unauthenticated
        '404':
          description: User not found
  /users/me/blocked/{blockedID}:
    delete:
      tags:
        - Followers
      summary: Ublock user
      description: unblocks a previously blocked user
      operationId: unbanUser
      parameters:
        - in: path
          name: blockedID
          description: User to unblock
          required: true
          schema:
            $ref: '#/components/schemas/ID'
      responses:
        '204':
          description: Successful operation
        '400':
          description: Bad auth token, bad userID or trying to unblock yourself
        '401':
          description: Unauthenticated
        '404':
          description: User not found or not blocked
  /users/me/following:
    get:
      tags:
        - Followers
      summary: List follows
      description: See who you're following
      operationId: getFollowing
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ID'
        '400':
          description: Bad auth token
        '401':
          description: Unauthenticated
        '403':
          description: "Forbidden: cannot view somebody else's follows"
    post:
      tags:
        - Followers
      summary: Follow user
      description: Follows a user. On success returns the followed user's ID
      operationId: followUser
      requestBody:
        description: user ID to follow
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ID'
      responses:
        '201':
          description: Successful operation
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ID'
        '400':
          description: bad auth token, bad userID, already following or trying to follow yourself
        '401':
          description: Unauthenticated
        '403':
          description: 'Forbidden: user blocked you!'
        '404':
          description: User not found
  /users/me/following/{followingID}:
    delete:
      tags:
        - Followers
      summary: Unfollow user
      description: Unfollows a user
      operationId: unfollowUser
      parameters:
        - in: path
          name: followingID
          description: User to unfollow
          required: true
          schema:
            $ref: '#/components/schemas/ID'
      responses:
        '204':
          description: Successful operation
        '400':
          description: Bad auth token, bad userID or trying to unfollow yourself
        '401':
          description: Unauthenticated
        '404':
          description: User not found or not following
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    ID:
      type: integer
      format: uint64
      description: ID represents an integer used as an identifier (such as postID and commentID)
      minimum: 0
      example: 32156
...
    Username:
      type: string
      example: Maria
      pattern: 'ˆ.*?$'
      description: Case-insensitive username
      minLength: 3
      maxLength: 40
    Password:
      type: string
      example: secret1234
      description: A user's password
      minLength: 8
      maxLength: 255
    UserCredentials:
      type: object
      description: Login credentials from a login form
      properties: 
        username:
          $ref: '#/components/schemas/Username'
        password:
          $ref: '#/components/schemas/Password'
    RegistrationRequest:
      type: object
      description: Registration info from a registration form
      properties:
        username:
          $ref: '#/components/schemas/Username'
        password:
          $ref: '#/components/schemas/Password'
    AccessToken:
      type: string
      description: A user's access token (opaque)
...
    UserList:
      type: array
      description: A list of users
      items:
        $ref: '#/components/schemas/ID'
...
security:
  - bearerAuth: []